requirements.txt
neural-renderer-pytorch
numpy
opencv-python
pyopengl
pyrender
scikit-image
scipy==1.0.0
tensorboard
chumpy
smplx
spacepy
# torch==1.1.0
torchgeometry
# torchvision
tqdm
trimesh
